Copy licenses over from https://github.com/ctz/rustls
The README mentions that ct-logs is licensed by Apache 2.0, ISC,
and MIT, just like rustls, but the actual license files are not
in the repo. This PR copies the text over from rustls in order
to make sure the explicit text is present in the published
artifacts.

In addition, this bumps the version of the crate in order to make
sure the artifacts on crates.io include these license files. If
this PR is accepted, could a new version of ct-logs be published
along side of it?

Closes #4