Update logs
This updates the latest logs. This adds the following logs:

* Google 'Xenon2019' log
* Google 'Xenon2020' log
* Google 'Xenon2021' log
* Google 'Xenon2022' log

This disqualifies a number of logs:

* Symantec log
* Symantec 'Vega' log
* Symantec 'Sirius' log

Which were all disqualified on Fri Feb 15 16:00:00 2019. This was
announced in:

https://groups.google.com/a/chromium.org/forum/#!msg/ct-policy/G2-s4YHeEEw/V9-i_ZscGQAJ

This also extends the build.py to print out disqualified logs
(which helped me to compare the generated code to the logs found on:

https://www.certificate-transparency.org/known-logs